+++ projects/angular-esri-components/src/lib/widgets/layers-toggle/sub-layers-toggle/sub-layers-toggle.component.spec.ts
@@ -1,0 +1,24 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { SubLayersToggleComponent } from './sub-layers-toggle.component';
+
+describe('SubLayersToggleComponent', () => {
+  let component: SubLayersToggleComponent;
+  let fixture: ComponentFixture<SubLayersToggleComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [SubLayersToggleComponent]
+    }).compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(SubLayersToggleComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});