+++ projects/angular-esri-components/src/testing/custom-widget.directive.ts
@@ -1,0 +1,8 @@
+import { Directive, Input } from '@angular/core';
+
+@Directive({
+  selector: '[esriCustomWidget], [esri-customWidget]'
+})
+export class MockCustomWidgetDirective {
+  @Input() position: string;
+}