+++ projects/angular-esri-components/src/testing/esri-module-provider.ts
@@ -1,0 +1,3 @@
+export class MockEsriModuleProvider {
+  require = jasmine.createSpy('require');
+}