+++ src/lib/index.ts
@@ -1,3 +1,0 @@
-export * from './esri-map/index';
-export * from './core/index';
-export * from './widgets/index';